Move Stringify from Support to Utils and simplify headers.
Support vs. Utils is a weird distinction and we should probably get rid
of it. I moved the biggest piece of code, Stringify, to Utils, and got
rid of a few headers. Things are simpler now.
