+++ Sources/Host/Linux/ARM/PTraceARM.cpp
@@ -12,15 +12,15 @@
 #include "DebugServer2/HardwareBreakpointManager.h"
 #include "DebugServer2/Host/Linux/ExtraWrappers.h"
 #include "DebugServer2/Host/Platform.h"
-#include "DebugServer2/Support/Stringify.h"
 #include "DebugServer2/Utils/Log.h"
+#include "DebugServer2/Utils/Stringify.h"
 
 #include <asm/ptrace.h>
 #include <elf.h>
 #include <sys/ptrace.h>
 #include <sys/uio.h>
 
-using ds2::Support::Stringify;
+using ds2::Utils::Stringify;
 
 #define super ds2::Host::POSIX::PTrace
